MODTLR-48 conflicts resolving, refactoring
