+++ src/main/java/org/folio/service/KafkaEventHandler.java
@@ -4,7 +4,5 @@ import org.folio.support.KafkaEvent;
 import org.springframework.messaging.MessageHeaders;
 
 public interface KafkaEventHandler<T> {
-  void handle(KafkaEvent<T> event);
-  void handleUserGroupCreatingEvent(KafkaEvent event, MessageHeaders messageHeaders);
-  void handleUserGroupUpdatingEvent(KafkaEvent event, MessageHeaders messageHeaders);
+  void handle(KafkaEvent<T> event, MessageHeaders messageHeaders);
 }
